Extracted stateless fragment XIC/match helpers to TopFragmentExtractor
* Moved ExtractTopNFragmentXics, ExtractFragmentXics, and CountTop6Matches (plus the CAL_TOP_N_FRAGMENTS constant) out of AbstractScoringTask into a new stateless OspreySharp.Scoring.TopFragmentExtractor; arithmetic and the stable-sort tie-break are verbatim
* Rewired the call sites in AbstractScoringTask and Calibrator; consolidated a duplicated doc tag; left s_calXcorrScorer in place (deferred)
* Stage B of the AbstractScoringTask decomposition; AbstractScoringTask -231 lines, byte-parity + perf gates green

See ai/todos/active/TODO-20260611_ospreysharp_decouple_abstractscoring.md
